tests/submissions/anssik/battery-discharging.html

<!DOCTYPE html>
<html>
  <head>
  <title>Battery Status API Test Suite</title>
  <script src="http://www.w3c-test.org/resources/testharness.js"></script>
  <link rel='stylesheet' href='http://w3c-test.org/resources/testharness.css' media='all'/>
  </head>
  <body>
    <h1>Description</h1>
    <p>
      The battery is discharging.
    </p>
    <h2>Preconditions</h2>
    <ol>
      <li>
        The device is unplugged from the charger before this test case is run.
      </li>
      <li>
        The battery must neither be empty or full, nor reach empty or full capacity during the test.
      </li>
    </ol>
    <p>
      The highest prime number discovered so far is:  <output id="prime"></output>
    </p>
    <div id="log"></div>
    <script>

    (function() {

      setup({ timeout: 5*60*1000 });
      
      test(function() {
        assert_false(navigator.battery.charging);
      }, 'The charging attribute must be set to false if the battery is discharging.');
      
      test(function() {
        assert_true(navigator.battery.chargingTime === Infinity);
      }, 'The chargingTime attribute must be set to the value positive Infinity if the battery is discharging');
      
      test(function() {
        assert_true(navigator.battery.dischargingTime < Infinity);
      }, 'The dischargingTime attribute must be set to the time remaining in seconds until the system\'s battery is completely discharged.');
      
      test(function() {
        assert_true(0 <= navigator.battery.level && navigator.battery.level <= 1.0);
      }, 'The level attribute must be set to the current battery level scaled from 0 to 1.0');
      
      var onlevelchange_test = async_test('When the battery\'s level changes, must decrease the level attribute\'s value and fire a levelchange event.');
      var battery_level = navigator.battery.level;
      
      // compute primes to deplete the battery faster
      var w = new Worker('prime.js');
      w.postMessage('compute');
      w.onmessage = function (e) {
        document.querySelector('#prime').textContent = e.data;
      }
      
      navigator.battery.onlevelchange = onlevelchange_test.step_func(function (e) {
        assert_true(navigator.battery.level < battery_level, 'The value of the level attribute must decrease, if the battery is discharging.');
        onlevelchange_test.done();
        w.terminate();
      });
      
    })();
    </script>
  </body>
</html>
